+++ client/src/App.js
@@ -2,9 +2,9 @@ import React from 'react';
 import './App.css';
 import { Route, Switch } from 'react-router-dom';
 import { LandingPage } from './components/LandingPage/LandingPage';
-import {UserOverviewEdit} from './components/UserOverviewEdit/UserOverviewEdit';
-import { UserOverviewView } from './components/UserOverviewEdit/UserOverviewView';
-import OverviewRouter from './components/UserOverviewEdit/OverviewRouter';
+import {UserOverviewEdit} from './components/UserOverview/UserOverviewEdit';
+import { UserOverviewView } from './components/UserOverview/UserOverviewView';
+import OverviewRouter from './components/UserOverview/OverviewRouter';
 
 
 function App() {
